Allow protocol and cql versions to be configured via driver path
In order to be used with Cassandra 1.2.x, the `ProtocolVersion` value
has to be set to 1 (gocql defaults to 2). This patch makes the migrate
code parse additional options out of the url.

Tests have been added for this new behavior, as well as to verify that
omitting the optional parameters preserves the defaults as is.