Use an empty string as parameter in ioutil.TempDir
The first parameter to ioutil.TempDir is the location where the
temporary directory will be created. An empty string tells Go to use
os.TempDir, otherwise it will try to create in the provided directory.

Using / works like a charm on Docker when running tests as root, but
won't work locally.