use timestamp (to seconds precision) for versions
Depends on #94. That'll need to be reviewed and merged first.

Using timestamps instead of an incrementing counter reduces the chances of conflicts and invalid migrations states after merging in another branch of code. Since people tend to work on orthogonal parts of the codebase, having to coordinate over incrementing an integer tends to cause merge issues late in the process of putting their code all together.

Along the way, we have to handle the fact that the original MySQL and PostgreSQL tables for migration versions used 32-bit integers instead of 64-bit integers. We inspect the column types of their migration tables and upgrade them non-destructively to 64-integers at the same time we check that the migration table exists. The other databases seem to already have the correct table types.

We also have to introduce a Clock interface for testing the new code
well. A Clock is just a handy way to control for time in tests. Since
we're constrained by the public API, the Clock instance is a global we
swap out during tests. This is not ideal, but usable.