feat(mysql): try creating database if not exist
**Background:** Our scenario is that we'd like to be able to create schema from scratch
for a brand new staging environment. This new staging environment
doesn't have the database created.

**Problem:** Having a first migration doing `CREATE DATABASE xxx`
doesn't work because a connection to `mysql://tcp(a.b.c.d)/xxx` fails
when database driver is trying to switch to the database `xxx`.

**Proposed solution:** When `Open(...)`ing a MySQL `Driver`, we always
try to create the database if it doesn't exist and _then_ we reestablish
a new database connection and ask the MySQL `database/sql/driver.Driver`
to make the switch to the database.